+++ src/lib/fonts.ts
@@ -11,21 +11,21 @@ const inter = Inter({
 const dystopian = localFont({
   src: [
     {
-      path: "../public/fonts/Dystopian-Light.woff2",
+      path: "../../public/fonts/Dystopian-Light.woff2",
       weight: "300",
       style: "normal",
     },
     {
-      path: "../public/fonts/Dystopian-Regular.woff2",
+      path: "../../public/fonts/Dystopian-Regular.woff2",
       style: "normal",
     },
     {
-      path: "../public/fonts/Dystopian-Bold.woff2",
+      path: "../../public/fonts/Dystopian-Bold.woff2",
       weight: "600",
       style: "normal",
     },
     {
-      path: "../public/fonts/Dystopian-Black.woff2",
+      path: "../../public/fonts/Dystopian-Black.woff2",
       weight: "700",
       style: "normal",
     },
